Code un-factoring and added Text Input element
Added a Text Input element that has all the features of a
Text Label but can also store user input text when clicked on.
It also uses interactivity states to give the user visual signs
of what state it's in (active, unactive, etc.).
Code un-factoring means I removed the classes, polymorphism, etc.
I changed everything to be structs and related functions in the
C-style. I did use template code for a function. This is because
trying to use classes was making things more complicated for me
to code, extend and debug. I will only refactor after everything
has been implemented.